optimize display texture drawing
