+++ libm8/src/constants/9_9.bmp
